Less logic in preprare_request_for_build
Logic in prepare_request_for_build and relevant helper methods in
build.py were moved to utils.py so it can be reused in
build_merge_index_image.py and potential future build modules
