+++ iib/workers/tasks/utils.py
@@ -19,9 +19,139 @@ from operator_manifest.operator import ImageName
 
 from iib.exceptions import IIBError
 from iib.workers.config import get_worker_config
+from iib.workers.api_utils import set_request_state
 
 log = logging.getLogger(__name__)
 dogpile_cache_region = create_dogpile_region()
+
+
+def get_binary_image_from_config(ocp_version, distribution_scope, binary_image_config={}):
+    """
+    Determine the binary image to be used to build the index image.
+
+    :param str ocp_version: the ocp_version label value of the index image.
+    :param str distribution_scope: the distribution_scope label value of the index image.
+    :param dict binary_image_config: the dict of config required to identify the appropriate
+        ``binary_image`` to use.
+    :return: pull specification of the binary_image to be used for this build.
+    :rtype: str
+    :raises IIBError: when the config value for the ocp_version and distribution_scope is missing.
+    """
+    binary_image = binary_image_config.get(distribution_scope, {}).get(ocp_version, None)
+    if not binary_image:
+        raise IIBError(
+            'IIB does not have a configured binary_image for'
+            f' distribution_scope : {distribution_scope} and ocp_version: {ocp_version}.'
+            ' Please specify a binary_image value in the request.'
+        )
+
+    return binary_image
+
+
+class RequestConfig:
+    """Request config abstract class.
+
+    :param str _binary_image:  the pull specification of the container image
+                          where the opm binary gets copied from.
+    :param str distribution_scope: the scope for distribution
+        of the index image, defaults to ``None``.
+    :param str source_from_index: the pull specification of the container image
+        containing the index that will be used
+        as a base of the merged index image.
+    :param str target_index: the pull specification of the container image
+        containing the index whose new data will be added
+        to the merged index image.
+    :param dict binary_image_config: the dict of config required to
+        identify the appropriate ``binary_image`` to use.
+    """
+
+    _attrs = [
+        "_binary_image",
+        "distribution_scope",
+        "source_from_index",
+        "target_index",
+        "binary_image_config",
+    ]
+    __slots__ = _attrs
+
+    def __init__(self, **kwargs):
+        """
+        Request config __init__.
+
+        Do not use this directly, use subclasses instead.
+        :Keyword Arguments:
+            See `_attrs` to check accepted keyword arguments.
+        """
+        for key in self.__slots__:
+            setattr(self, key, None)
+        for key, val in kwargs.items():
+            setattr(self, key, kwargs[key])
+
+    def __eq__(self, other):
+        if type(self) == type(other) and [getattr(self, x) for x in self.__slots__] == [
+            getattr(self, x) for x in self.__slots__
+        ]:
+            return True
+        return False
+
+    def binary_image(self, index_info, distribution_scope):
+        """Get binary image based on self configuration, index image info and distribution scope."""
+        if not self._binary_image:
+            binary_image_ocp_version = index_info['ocp_version']
+            return get_binary_image_from_config(
+                binary_image_ocp_version, distribution_scope, self.binary_image_config
+            )
+        return self._binary_image
+
+
+class RequestConfigAddRm(RequestConfig):
+    """Request config for add and remove operations.
+
+    :param str overwrite_from_index_token: the token used for overwriting the input
+        ``from_index`` image. This is required for
+        non-privileged users to use ``overwrite_from_index``.
+        The format of the token must be
+        in the format "user:password".
+    :param str from_index: the pull specification of the container image
+        containing the index that the index image build
+        will be based from.
+    :param list add_arches: the list of arches to build in addition to the
+        arches ``from_index`` is currently built for;
+        if ``from_index`` is ``None``, then this is used as the list of arches
+        to build the index image for
+    :param list bundles: the list of bundles to create the
+        bundle mapping on the request
+    """
+
+    _attrs = RequestConfig._attrs + [
+        "overwrite_from_index_token",
+        "from_index",
+        "add_arches",
+        "bundles",
+    ]
+    __slots__ = _attrs
+
+
+class RequestConfigMerge(RequestConfig):
+    """Request config for merge operation.
+
+    :param str overwrite_target_index_token:  auth token used to pus index image
+        when overwrite is set.
+    :param str source_from_index: the pull specification of the container image
+        containing the index that will be used
+        as a base of the merged index image.
+    :param str target_index: the pull specification of the container image
+        containing the index whose new data will be added
+        to the merged index image.
+    """
+
+    _attrs = RequestConfig._attrs + [
+        "source_from_index",
+        "target_index",
+        "overwrite_target_index_token",
+    ]
+
+    __slots__ = _attrs
 
 
 def deprecate_bundles(
@@ -184,9 +314,7 @@ def get_image_labels(pull_spec):
     return skopeo_inspect(full_pull_spec, '--config').get('config', {}).get('Labels', {})
 
 
-def retry(
-    attempts=get_worker_config().iib_total_attempts, wait_on=Exception, logger=None,
-):
+def retry(attempts=get_worker_config().iib_total_attempts, wait_on=Exception, logger=None):
     """
     Retry a section of code until success or max attempts are reached.
 
@@ -494,3 +622,243 @@ def chmod_recursively(dir_path, dir_mode, file_mode):
             if not os.path.exists(file_path) or os.path.islink(file_path):
                 continue
             os.chmod(file_path, file_mode)
+
+
+def gather_index_image_arches(build_request_config, index_image_infos):
+    """Gather architectures from build_request_config and provided index image.
+
+    :param RequestConfig build_request_config: build request configuration
+    :param dict index_image_infos: dict with index image infos returned
+        by `get_all_index_images_info`
+    :return: set of architecture of all index images
+    :rtype: set
+    """
+    arches = set(build_request_config.add_arches or [])
+    for info in index_image_infos.values():
+        arches |= set(info['arches'])
+
+    if not arches:
+        raise IIBError('No arches were provided to build the index image')
+    return arches
+
+
+def get_image_arches(pull_spec):
+    """
+    Get the architectures this image was built for.
+
+    :param str pull_spec: the pull specification to a v2 manifest list
+    :return: a set of architectures of the container images contained in the manifest list
+    :rtype: set
+    :raises IIBError: if the pull specification is not a v2 manifest list
+    """
+    log.debug('Get the available arches for %s', pull_spec)
+    skopeo_raw = skopeo_inspect(f'docker://{pull_spec}', '--raw')
+    arches = set()
+    if skopeo_raw.get('mediaType') == 'application/vnd.docker.distribution.manifest.list.v2+json':
+        for manifest in skopeo_raw['manifests']:
+            arches.add(manifest['platform']['architecture'])
+    elif skopeo_raw.get('mediaType') == 'application/vnd.docker.distribution.manifest.v2+json':
+        skopeo_out = skopeo_inspect(f'docker://{pull_spec}', '--config')
+        arches.add(skopeo_out['architecture'])
+    else:
+        raise IIBError(
+            f'The pull specification of {pull_spec} is neither a v2 manifest list nor a v2 manifest'
+        )
+
+    return arches
+
+
+def get_index_image_info(overwrite_from_index_token, from_index=None, default_ocp_version='v4.5'):
+    """Get arches, resolved pull specification and ocp_version for the index image.
+
+    :param str overwrite_from_index_token: the token used for overwriting the input
+        ``from_index`` image. This is required for non-privileged users to use
+        ``overwrite_from_index``. The format of the token must be in the format "user:password".
+    :param str from_index: the pull specification of the index image to be resolved.
+    :param str default_ocp_version: default ocp_version to use if index image pull_spec is absent.
+    :return: dictionary of resolved index image pull spec, set of arches, default ocp_version and
+        resolved_distribution_scope
+    :rtype: dict
+    """
+    result = {
+        'resolved_from_index': None,
+        'ocp_version': default_ocp_version,
+        'arches': set(),
+        'resolved_distribution_scope': 'prod',
+    }
+    if not from_index:
+        return result
+
+    with set_registry_token(overwrite_from_index_token, from_index):
+        from_index_resolved = get_resolved_image(from_index)
+        result['arches'] = get_image_arches(from_index_resolved)
+        result['ocp_version'] = (
+            get_image_label(from_index_resolved, 'com.redhat.index.delivery.version') or 'v4.5'
+        )
+        result['resolved_distribution_scope'] = (
+            get_image_label(from_index_resolved, 'com.redhat.index.delivery.distribution_scope')
+            or 'prod'
+        )
+        result['resolved_from_index'] = from_index_resolved
+    return result
+
+
+def get_all_index_images_info(build_request_config, index_version_map):
+    """Get image info of all images in version map.
+
+    :param RequestConfig build_request_config: build request configuration
+    :param list index_version_map: list of tuples with (index_name, index_ocp_version)
+    :return: dictionary with inex image information obtained from `get_index_image_info`
+    :rtype: dict
+    """
+    infos = {}
+    for (index, version) in index_version_map:
+        if not hasattr(build_request_config, index):
+            from_index = None
+        else:
+            from_index = getattr(build_request_config, index)
+
+        infos[index] = get_index_image_info(
+            build_request_config.overwrite_from_index_token
+            if hasattr(build_request_config, 'overwrite_from_index_token')
+            else build_request_config.overwrite_target_index_token,
+            from_index=from_index,
+            default_ocp_version=version,
+        )
+    return infos
+
+
+def get_image_label(pull_spec, label):
+    """
+    Get a specific label from the container image.
+
+    :param str pull_spec: pull spec of the image
+    :param str label: the label to get
+    :return: the label on the container image or None
+    :rtype: str
+    """
+    log.debug('Getting the label of %s from %s', label, pull_spec)
+    return get_image_labels(pull_spec).get(label)
+
+
+def verify_labels(bundles):
+    """
+    Verify that the required labels are set on the input bundles.
+
+    :param list bundles: a list of strings representing the pull specifications of the bundles to
+        add to the index image being built.
+    :raises IIBError: if one of the bundles does not have the correct label value.
+    """
+    conf = get_worker_config()
+    if not conf['iib_required_labels']:
+        return
+
+    for bundle in bundles:
+        labels = get_image_labels(bundle)
+        for label, value in conf['iib_required_labels'].items():
+            if labels.get(label) != value:
+                raise IIBError(f'The bundle {bundle} does not have the label {label}={value}')
+
+
+def _validate_distribution_scope(resolved_distribution_scope, distribution_scope):
+    """
+    Validate distribution scope is allowed to be updated.
+
+    :param str resolved_distribution_scope: the distribution_scope that the index is for.
+    :param str distribution_scope: the distribution scope that has been requested for
+        the index image.
+    :return: the valid distribution scope
+    :rtype: str
+    :raises IIBError: if the ``resolved_distribution_scope`` is of lesser scope than
+        ``distribution_scope``
+    """
+    if not distribution_scope:
+        return resolved_distribution_scope
+
+    scopes = ["dev", "stage", "prod"]
+    # Make sure the request isn't regressing the distribution scope
+    if scopes.index(distribution_scope) > scopes.index(resolved_distribution_scope):
+        raise IIBError(
+            f'Cannot set "distribution_scope" to {distribution_scope} because from index is'
+            f' already set to {resolved_distribution_scope}'
+        )
+    return distribution_scope
+
+
+def prepare_request_for_build(request_id, build_request_config):
+    """Prepare the request for the index image build.
+
+    All information that was retrieved and/or calculated for the next steps in the build are
+    returned as a dictionary.
+    This function was created so that code didn't need to be duplicated for the ``add`` and ``rm``
+    request types.
+    :param RequestConfig build_request_config: build request configuration
+    :return: a dictionary with the keys: arches, binary_image_resolved, from_index_resolved, and
+    ocp_version.
+    :rtype: dict
+    :raises IIBError: if the container image resolution fails or the architectures couldn't be
+    detected.
+    """
+    bundles = None
+    if hasattr(build_request_config, "bundles"):
+        bundles = build_request_config.bundles
+
+    if bundles is None:
+        bundles = []
+
+    set_request_state(request_id, 'in_progress', 'Resolving the container images')
+
+    # Use v4.5 as default version
+    index_info = get_all_index_images_info(
+        build_request_config,
+        [("from_index", "v4.5"), ("source_from_index", "v4.5"), ("target_index", "v4.6")],
+    )
+    arches = gather_index_image_arches(build_request_config, index_info)
+    if not arches:
+        raise IIBError('No arches were provided to build the index image')
+
+    arches_str = ', '.join(sorted(arches))
+    log.debug('Set to build the index image for the following arches: %s', arches_str)
+
+    # Use the distribution_scope of the from_index as the resolved distribution scope for `Add`,
+    # and 'Rm' requests, but use the distribution_scope of the target_index as the resolved
+    # distribution scope for `merge-index-image` requests.
+    resolved_distribution_scope = index_info['from_index']['resolved_distribution_scope']
+    if build_request_config.source_from_index:
+        resolved_distribution_scope = index_info['target_index']['resolved_distribution_scope']
+
+    distribution_scope = _validate_distribution_scope(
+        resolved_distribution_scope, build_request_config.distribution_scope
+    )
+
+    binary_image = build_request_config.binary_image(index_info['from_index'], distribution_scope)
+
+    binary_image_resolved = get_resolved_image(binary_image)
+    binary_image_arches = get_image_arches(binary_image_resolved)
+
+    if not arches.issubset(binary_image_arches):
+        raise IIBError(
+            'The binary image is not available for the following arches: {}'.format(
+                ', '.join(sorted(arches - binary_image_arches))
+            )
+        )
+
+    bundle_mapping = {}
+    for bundle in bundles:
+        operator = get_image_label(bundle, 'operators.operatorframework.io.bundle.package.v1')
+        if operator:
+            bundle_mapping.setdefault(operator, []).append(bundle)
+
+    return {
+        'arches': arches,
+        'binary_image': binary_image,
+        'binary_image_resolved': binary_image_resolved,
+        'bundle_mapping': bundle_mapping,
+        'from_index_resolved': index_info["from_index"]['resolved_from_index'],
+        'ocp_version': index_info["from_index"]['ocp_version'],
+        'distribution_scope': distribution_scope,
+        'source_from_index_resolved': index_info['source_from_index']['resolved_from_index'],
+        'source_ocp_version': index_info['source_from_index']['ocp_version'],
+        'target_index_resolved': index_info['target_index']['resolved_from_index'],
+        'target_ocp_version': index_info['target_index']['ocp_version'],
+    }
