+++ tests/test_workers/test_tasks/test_build_regenerate_bundle.py
@@ -27,7 +27,7 @@ yaml = ruamel.yaml.YAML()
 @mock.patch('iib.workers.tasks.build_regenerate_bundle.get_resolved_image')
 @mock.patch('iib.workers.tasks.build_regenerate_bundle.podman_pull')
 @mock.patch('iib.workers.tasks.build_regenerate_bundle.tempfile.TemporaryDirectory')
-@mock.patch('iib.workers.tasks.build_regenerate_bundle._get_image_arches')
+@mock.patch('iib.workers.tasks.build_regenerate_bundle.get_image_arches')
 @mock.patch('iib.workers.tasks.build_regenerate_bundle._copy_files_from_image')
 @mock.patch('iib.workers.tasks.build_regenerate_bundle._adjust_operator_bundle')
 @mock.patch('iib.workers.tasks.build_regenerate_bundle._build_image')
