+++ tox.ini
@@ -20,6 +20,7 @@ setenv =
 	IIB_TESTING=true
 pytest_command =
     pytest -vv \
+        --capture=sys\
         --cov-config .coveragerc --cov=iib --cov-report term \
         --cov-report xml --cov-report html {posargs}
